+++ src/Transfer/Attribute/EnumPropertyTypeAttribute.php
@@ -39,9 +39,4 @@ final readonly class EnumPropertyTypeAttribute implements PropertyTypeAttributeI
     {
         return $data?->value;
     }
-
-    public function getInitialValue(): null
-    {
-        return null;
-    }
 }
